requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Celery
celery==5.3.6
redis==5.0.1

# Storage
google-cloud-storage==2.14.0
boto3==1.34.34

# OCR & PDF
paddleocr==2.7.0.3
paddlepaddle==2.6.0
PyMuPDF==1.23.21
Pillow==10.2.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
